requirements.txt
aiofiles==24.1.0
fastapi==0.115.12
line_bot_sdk==3.16.3
rich==14.0.0
uvicorn==0.34.2
